+++ Lost Histories/ItemConsumable.h
@@ -9,7 +9,7 @@ class ItemConsumable : public Item
 private:
 public:
 	ItemConsumable();
-	ItemConsumable(string nName, string nDesc, int nRarity, string nType, int nAmount);
+	ItemConsumable(string nName, string nDesc, int nRarity, string nType, double nAmount);
 	string toString() override;
 };
